+++ tests/test_classification_krige.py
@@ -3,14 +3,13 @@ import pytest
 
 import numpy as np
 
-from pykrige.rk import RegressionKriging
+from pykrige.ck import ClassificationKriging
 
 try:
-    from sklearn.svm import SVR
+    from sklearn.svm import SVC
     from sklearn.datasets import fetch_california_housing
-    from sklearn.linear_model import ElasticNet, Lasso
-    from sklearn.ensemble import RandomForestRegressor
-    from sklearn.linear_model import LinearRegression
+    from sklearn.ensemble import RandomForestClassifier
+    from sklearn.preprocessing import KBinsDiscretizer
     from sklearn.model_selection import train_test_split
 
     SKLEARN_INSTALLED = True
@@ -21,17 +20,14 @@ except ImportError:
 def _methods():
     krige_methods = ["ordinary", "universal"]
     ml_methods = [
-        SVR(C=0.01, gamma="auto"),
-        RandomForestRegressor(min_samples_split=5, n_estimators=50),
-        LinearRegression(),
-        Lasso(),
-        ElasticNet(),
+        SVC(C=0.01, gamma="auto", probability=True),
+        RandomForestClassifier(n_estimators=50),
     ]
     return product(ml_methods, krige_methods)
 
 
 @pytest.mark.skipif(not SKLEARN_INSTALLED, reason="requires scikit-learn")
-def test_regression_krige():
+def test_classification_krige():
     np.random.seed(1)
     x = np.linspace(-1.0, 1.0, 100)
     # create a feature matrix with 5 features
@@ -51,20 +47,23 @@ def test_regression_krige():
     lat = np.linspace(-90.0, 90.0, 10)
     lon_lat = np.array(list(product(lon, lat)))
 
+    discretizer = KBinsDiscretizer(encode="ordinal")
+    y = discretizer.fit_transform(y.reshape(-1, 1))
+
     X_train, X_test, y_train, y_test, lon_lat_train, lon_lat_test = train_test_split(
         X, y, lon_lat, train_size=0.7, random_state=10
     )
 
     for ml_model, krige_method in _methods():
-        reg_kr_model = RegressionKriging(
-            regression_model=ml_model, method=krige_method, n_closest_points=2
+        class_model = ClassificationKriging(
+            classification_model=ml_model, method=krige_method, n_closest_points=2
         )
-        reg_kr_model.fit(X_train, lon_lat_train, y_train)
-        assert reg_kr_model.score(X_test, lon_lat_test, y_test) > 0.25
+        class_model.fit(X_train, lon_lat_train, y_train)
+        assert class_model.score(X_test, lon_lat_test, y_test) > 0.25
 
 
 @pytest.mark.skipif(not SKLEARN_INSTALLED, reason="requires scikit-learn")
-def test_krige_housing():
+def test_krige_classification_housing():
     import ssl
     import urllib
 
@@ -83,6 +82,8 @@ def test_krige_housing():
     p = housing["data"][:1000, :-2]
     x = housing["data"][:1000, -2:]
     target = housing["target"][:1000]
+    discretizer = KBinsDiscretizer(encode="ordinal")
+    target = discretizer.fit_transform(target.reshape(-1, 1))
 
     p_train, p_test, y_train, y_test, x_train, x_test = train_test_split(
         p, target, x, train_size=0.7, random_state=10
@@ -90,11 +91,11 @@ def test_krige_housing():
 
     for ml_model, krige_method in _methods():
 
-        reg_kr_model = RegressionKriging(
-            regression_model=ml_model, method=krige_method, n_closest_points=2
+        class_model = ClassificationKriging(
+            classification_model=ml_model, method=krige_method, n_closest_points=2
         )
-        reg_kr_model.fit(p_train, x_train, y_train)
+        class_model.fit(p_train, x_train, y_train)
         if krige_method == "ordinary":
-            assert reg_kr_model.score(p_test, x_test, y_test) > 0.5
+            assert class_model.score(p_test, x_test, y_test) > 0.5
         else:
-            assert reg_kr_model.score(p_test, x_test, y_test) > 0.0
+            assert class_model.score(p_test, x_test, y_test) > 0.0